Extract remaining settings sections
