+++ src/lib/components/SettingsPanel.svelte
@@ -1,41 +1,4 @@
 <script lang="ts">
-  import {
-    settings,
-    setStartupTarget,
-    updateSetting,
-  } from "$lib/stores/settings";
-  import {
-    sidebarLayout,
-    setRailSide,
-    type Side,
-  } from "$lib/stores/sidebarLayout";
-  import { open } from "@tauri-apps/plugin-dialog";
-  import { THEME_DEFINITIONS } from "$lib/themes";
-  import {
-    quitApp,
-  } from "$lib/tauri";
-  import { commands } from "$lib/bindings";
-  import type {
-    ExternalTool,
-    ExternalToolSurface,
-    ExternalToolWebEmbedder,
-    IntegrationDetection,
-    McpHostConfigPreview,
-    McpStatus,
-    OnPaneCloseMode,
-    StartupTarget,
-    UpdateChannel,
-    WorktreeCleanupMode,
-    WorktreeDefaultBase,
-    WorktreeProvider,
-    WorktrunkDetection,
-  } from "$lib/bindings";
-  import { updateStatus, runManualCheck, performInstall } from "$lib/stores/updater";
-  import { activeSession } from "$lib/stores/sessions";
-  import { settingsFocus } from "$lib/stores/settingsFocus";
-  import { previewExternalToolConfig, type RenderedExternalTool } from "$lib/tauri";
-  import { getVersion } from "@tauri-apps/api/app";
-  import { onMount } from "svelte";
   import Settings from "@lucide/svelte/icons/settings";
   import FolderTree from "@lucide/svelte/icons/folder-tree";
   import TerminalIcon from "@lucide/svelte/icons/terminal";
@@ -48,14 +11,19 @@
   import ClipboardList from "@lucide/svelte/icons/clipboard-list";
   import FlaskConical from "@lucide/svelte/icons/flask-conical";
   import X from "@lucide/svelte/icons/x";
-  import SettingsKanbanSection from "$lib/components/settings/SettingsKanbanSection.svelte";
-  import SettingsNotesSection from "$lib/components/settings/SettingsNotesSection.svelte";
-  import SettingsKeyboardSection from "$lib/components/settings/SettingsKeyboardSection.svelte";
-  import SettingsExperimentsSection from "$lib/components/settings/SettingsExperimentsSection.svelte";
-  import SettingsNotificationsSection from "$lib/components/settings/SettingsNotificationsSection.svelte";
   import SettingsAdvancedSection from "$lib/components/settings/SettingsAdvancedSection.svelte";
-  import SettingsTerminalSection from "$lib/components/settings/SettingsTerminalSection.svelte";
   import SettingsAgentsSection from "$lib/components/settings/SettingsAgentsSection.svelte";
+  import SettingsExperimentsSection from "$lib/components/settings/SettingsExperimentsSection.svelte";
+  import SettingsExternalToolsSection from "$lib/components/settings/SettingsExternalToolsSection.svelte";
+  import SettingsGeneralSection from "$lib/components/settings/SettingsGeneralSection.svelte";
+  import SettingsIntegrationsSection from "$lib/components/settings/SettingsIntegrationsSection.svelte";
+  import SettingsKanbanSection from "$lib/components/settings/SettingsKanbanSection.svelte";
+  import SettingsKeyboardSection from "$lib/components/settings/SettingsKeyboardSection.svelte";
+  import SettingsNotesSection from "$lib/components/settings/SettingsNotesSection.svelte";
+  import SettingsNotificationsSection from "$lib/components/settings/SettingsNotificationsSection.svelte";
+  import SettingsSessionsSection from "$lib/components/settings/SettingsSessionsSection.svelte";
+  import SettingsTerminalSection from "$lib/components/settings/SettingsTerminalSection.svelte";
+  import { settingsFocus } from "$lib/stores/settingsFocus";
   import {
     normalizeSettingsCategoryId,
     type SettingsCategoryId,
@@ -76,43 +44,6 @@
     { id: "advanced", label: "Advanced", icon: Wrench },
   ];
 
-  const PANE_CLOSE_OPTIONS = [
-    { id: "kill", label: "Kill" },
-    { id: "detach", label: "Detach" },
-  ] as const;
-
-  const STARTUP_TARGET_OPTIONS: { id: StartupTarget; label: string }[] = [
-    { id: "restore", label: "Restore previous" },
-    { id: "sessionsSidebar", label: "Sessions sidebar" },
-    { id: "lastSession", label: "Last session" },
-    { id: "kanbanWide", label: "Kanban wide view" },
-    { id: "externalTool", label: "External tool" },
-    { id: "none", label: "None" },
-  ];
-
-  let selected = $state<SettingsCategoryId>("general");
-
-  let appVersion = $state<string>("…");
-  let repoRootDraft = $state("");
-  let expandedExternalToolId = $state<string | null>(null);
-  let externalToolPreviewById = $state<
-    Record<string, { loading: boolean; rendered: RenderedExternalTool | null; error: string | null }>
-  >({});
-  const externalToolRowKeys = new Map<string, string>();
-  let nextExternalToolRowKey = 0;
-  let wasVisible = false;
-  onMount(async () => {
-    try { appVersion = await getVersion(); } catch { appVersion = "unknown"; }
-  });
-
-  function describeError(reason: "network" | "signature-invalid" | "unknown"): string {
-    switch (reason) {
-      case "network": return "Couldn't reach the update server.";
-      case "signature-invalid": return "Update signature invalid. Download blocked.";
-      case "unknown": return "Update check failed.";
-    }
-  }
-
   interface Props {
     visible: boolean;
     onclose: () => void;
@@ -122,26 +53,40 @@
 
   let { visible, onclose, initialCategory = null, externalToolId = null }: Props = $props();
 
+  let selected = $state<SettingsCategoryId>("general");
+  let focusedExternalToolId = $state<string | null>(null);
+  let externalToolFocusToken = $state(0);
+  let nextExternalToolFocusToken = 0;
+  let wasVisible = false;
+
+  function focusExternalTool(id: string | null): void {
+    focusedExternalToolId = id;
+    externalToolFocusToken = ++nextExternalToolFocusToken;
+  }
+
   $effect(() => {
     const justOpened = visible && !wasVisible;
     wasVisible = visible;
     if (!visible) return;
+
     const focus = $settingsFocus;
     if (focus?.category) {
       selected = normalizeSettingsCategoryId(focus.category);
       if (focus.category === "externalTools" && "externalToolId" in focus) {
-        expandedExternalToolId = focus.externalToolId ?? null;
+        focusExternalTool(focus.externalToolId ?? null);
       }
       settingsFocus.set(null);
       return;
     }
+
     if (initialCategory) {
       selected = normalizeSettingsCategoryId(initialCategory);
       if (initialCategory === "externalTools") {
-        expandedExternalToolId = externalToolId ?? null;
+        focusExternalTool(externalToolId ?? null);
       }
       return;
     }
+
     if (justOpened) selected = "general";
   });
 
@@ -152,383 +97,6 @@
       onclose();
     }
   }
-
-  function externalTools(): ExternalTool[] {
-    return $settings.externalTools ?? [];
-  }
-
-  function externalToolRowKey(id: string): string {
-    let key = externalToolRowKeys.get(id);
-    if (!key) {
-      key = `external-tool-row-${++nextExternalToolRowKey}`;
-      externalToolRowKeys.set(id, key);
-    }
-    return key;
-  }
-
-  function retainExternalToolRowKey(previousId: string, nextId: string): void {
-    if (previousId === nextId) return;
-    const key = externalToolRowKeys.get(previousId);
-    if (!key) return;
-    externalToolRowKeys.delete(previousId);
-    externalToolRowKeys.set(nextId, key);
-  }
-
-  function pruneExternalToolRowKeys(tools: ExternalTool[]): void {
-    const ids = new Set(tools.map((tool) => tool.id));
-    for (const id of externalToolRowKeys.keys()) {
-      if (!ids.has(id)) externalToolRowKeys.delete(id);
-    }
-  }
-
-  function updateExternalTools(tools: ExternalTool[]): void {
-    pruneExternalToolRowKeys(tools);
-    updateSetting("externalTools", tools);
-  }
-
-  function updateExternalTool(id: string, patch: Partial<ExternalTool>): void {
-    const tools = externalTools();
-    let nextPatch = patch;
-    if (patch.id !== undefined) {
-      const normalizedId = patch.id.trim();
-      if (!normalizedId || tools.some((tool) => tool.id !== id && tool.id.trim() === normalizedId)) {
-        return;
-      }
-      nextPatch = { ...patch, id: normalizedId };
-    }
-    if (nextPatch.id !== undefined) {
-      retainExternalToolRowKey(id, nextPatch.id);
-      if (expandedExternalToolId === id) expandedExternalToolId = nextPatch.id;
-    }
-    updateExternalTools(tools.map((tool) => (tool.id === id ? { ...tool, ...nextPatch } : tool)));
-  }
-
-  function preferredPortFromInput(value: string): number | null {
-    const trimmed = value.trim();
-    if (!trimmed) return null;
-    const parsed = Number.parseInt(trimmed, 10);
-    if (Number.isNaN(parsed)) return null;
-    return Math.min(65535, Math.max(1, parsed));
-  }
-
-  function addExternalTool(surface: ExternalToolSurface): void {
-    const id = `tool-${Date.now()}`;
-    const tool: ExternalTool = {
-      id,
-      name: surface === "web" ? "New Web Tool" : "New Terminal Tool",
-      enabled: true,
-      surface,
-      commandTemplate: surface === "web" ? "server --port {{ port }}" : "command",
-      cwdTemplate: "{{ session.worktree_path }}",
-      requiresSession: true,
-      urlTemplate: surface === "web" ? "http://127.0.0.1:{{ port }}" : null,
-      preferredPort: surface === "web" ? 4966 : null,
-      webEmbedder: "webview",
-      keepWebviewAlive: false,
-    };
-    updateExternalTools([...externalTools(), tool]);
-    expandedExternalToolId = id;
-  }
-
-  function removeExternalTool(id: string): void {
-    updateExternalTools(externalTools().filter((tool) => tool.id !== id));
-    if (expandedExternalToolId === id) expandedExternalToolId = null;
-  }
-
-  async function previewTool(tool: ExternalTool): Promise<void> {
-    externalToolPreviewById = {
-      ...externalToolPreviewById,
-      [tool.id]: { loading: true, rendered: null, error: null },
-    };
-    try {
-      const rendered = await previewExternalToolConfig(
-        tool,
-        tool.requiresSession ? ($activeSession?.id ?? null) : null,
-        tool.surface === "web" ? (tool.preferredPort ?? 4966) : null,
-      );
-      externalToolPreviewById = {
-        ...externalToolPreviewById,
-        [tool.id]: { loading: false, rendered, error: null },
-      };
-    } catch (err) {
-      externalToolPreviewById = {
-        ...externalToolPreviewById,
-        [tool.id]: {
-          loading: false,
-          rendered: null,
-          error: err instanceof Error ? err.message : String(err),
-        },
-      };
-    }
-  }
-
-  async function browseGhBinary() {
-    const selected = await open({ directory: false, title: "Select gh (GitHub CLI) Binary" });
-    if (selected) updateSetting("ghBinaryPath", selected as string);
-  }
-
-  async function browseGitBinary() {
-    const selected = await open({ directory: false, title: "Select git Binary" });
-    if (selected) updateSetting("gitBinaryPath", selected as string);
-  }
-
-  async function browseWorktrunkBinary() {
-    const selected = await open({
-      directory: false,
-      title: "Select wt (worktrunk) Binary",
-    });
-    if (selected) updateSetting("worktrunkBinaryPath", selected as string);
-  }
-
-  function globalExternalTools(): ExternalTool[] {
-    return externalTools().filter((tool) => tool.enabled !== false && !(tool.requiresSession ?? false));
-  }
-
-  function updateStartupTarget(target: StartupTarget): void {
-    setStartupTarget(target);
-  }
-
-  let ghDetection = $state<IntegrationDetection | null>(null);
-  let gitDetection = $state<IntegrationDetection | null>(null);
-  let worktrunkDetection = $state<WorktrunkDetection | null>(null);
-  let mcpStatus = $state<McpStatus | null>(null);
-  // Per-host UI state, keyed by host id (`claudeDesktop`, `claudeCode`,
-  // `codex`). Each host gets its own preview blob + status message so a
-  // configure-Codex run doesn't blow away a preview-Claude-Code result.
-  let mcpPreviewByHost = $state<Record<string, McpHostConfigPreview | null>>({});
-  let mcpMessageByHost = $state<Record<string, string | null>>({});
-  let mcpErrorByHost = $state<Record<string, string | null>>({});
-  let mcpBusyByHost = $state<Record<string, "preview" | "configure" | null>>({});
-  let ghDetectionRun = 0;
-  let gitDetectionRun = 0;
-  let worktrunkDetectionRun = 0;
-  let mcpStatusRun = 0;
-
-  async function refreshGhDetection(run: number) {
-    try {
-      const result = await commands.cmdDetectGh();
-      if (run === ghDetectionRun) ghDetection = result;
-    } catch {
-      if (run === ghDetectionRun) ghDetection = { binaryPath: null, version: null };
-    }
-  }
-
-  async function refreshGitDetection(run: number) {
-    try {
-      const result = await commands.cmdDetectGit();
-      if (run === gitDetectionRun) gitDetection = result;
-    } catch {
-      if (run === gitDetectionRun) gitDetection = { binaryPath: null, version: null };
-    }
-  }
-
-  async function refreshWorktrunkDetection(run: number) {
-    try {
-      const result = await commands.cmdDetectWorktrunk(null);
-      if (run === worktrunkDetectionRun) worktrunkDetection = result;
-    } catch {
-      if (run === worktrunkDetectionRun) {
-        worktrunkDetection = { binaryPath: null, version: null, hasConfig: false };
-      }
-    }
-  }
-
-  async function refreshMcpStatus(run: number) {
-    try {
-      const result = await commands.cmdMcpStatus();
-      if (run === mcpStatusRun) mcpStatus = result;
-    } catch {
-      if (run === mcpStatusRun) mcpStatus = null;
-    }
-  }
-
-  $effect(() => {
-    const path = $settings.ghBinaryPath;
-    void path;
-    if (!visible || selected !== "integrations") {
-      ghDetectionRun += 1;
-      return;
-    }
-    const run = ++ghDetectionRun;
-    const timer = setTimeout(() => void refreshGhDetection(run), 250);
-    return () => clearTimeout(timer);
-  });
-
-  $effect(() => {
-    const path = $settings.gitBinaryPath;
-    void path;
-    if (!visible || selected !== "integrations") {
-      gitDetectionRun += 1;
-      return;
-    }
-    const run = ++gitDetectionRun;
-    const timer = setTimeout(() => void refreshGitDetection(run), 250);
-    return () => clearTimeout(timer);
-  });
-
-  $effect(() => {
-    const path = $settings.worktrunkBinaryPath;
-    void path;
-    if (!visible || selected !== "integrations") {
-      worktrunkDetectionRun += 1;
-      return;
-    }
-    const run = ++worktrunkDetectionRun;
-    const timer = setTimeout(() => void refreshWorktrunkDetection(run), 250);
-    return () => clearTimeout(timer);
-  });
-
-  $effect(() => {
-    const enabled = $settings.mcpEnabled;
-    void enabled;
-    if (!visible || selected !== "integrations") {
-      mcpStatusRun += 1;
-      return;
-    }
-    const run = ++mcpStatusRun;
-    const timer = setTimeout(() => void refreshMcpStatus(run), 250);
-    return () => clearTimeout(timer);
-  });
-
-  type McpHostIdT = "claudeDesktop" | "claudeCode" | "codex";
-
-  async function previewMcpHostConfig(hostId: McpHostIdT, label: string) {
-    mcpBusyByHost = { ...mcpBusyByHost, [hostId]: "preview" };
-    mcpErrorByHost = { ...mcpErrorByHost, [hostId]: null };
-    mcpMessageByHost = { ...mcpMessageByHost, [hostId]: null };
-    try {
-      const result = await commands.cmdPreviewMcpHostConfig(hostId);
-      if (result.status === "error") {
-        throw new Error(result.error);
-      }
-      mcpPreviewByHost = { ...mcpPreviewByHost, [hostId]: result.data };
-      mcpMessageByHost = {
-        ...mcpMessageByHost,
-        [hostId]: result.data.configured
-          ? `${label} is already configured.`
-          : "Preview ready.",
-      };
-    } catch (e) {
-      mcpPreviewByHost = { ...mcpPreviewByHost, [hostId]: null };
-      mcpErrorByHost = {
-        ...mcpErrorByHost,
-        [hostId]: e instanceof Error ? e.message : String(e),
-      };
-    } finally {
-      mcpBusyByHost = { ...mcpBusyByHost, [hostId]: null };
-    }
-  }
-
-  async function configureMcpHost(hostId: McpHostIdT, label: string) {
-    mcpBusyByHost = { ...mcpBusyByHost, [hostId]: "configure" };
-    mcpErrorByHost = { ...mcpErrorByHost, [hostId]: null };
-    mcpMessageByHost = { ...mcpMessageByHost, [hostId]: null };
-    try {
-      const result = await commands.cmdConfigureMcpHost(hostId);
-      if (result.status === "error") {
-        throw new Error(result.error);
-      }
-      mcpPreviewByHost = { ...mcpPreviewByHost, [hostId]: result.data };
-      mcpMessageByHost = {
-        ...mcpMessageByHost,
-        [hostId]: result.data.configured
-          ? `${label} was already configured.`
-          : `${label} configuration updated.`,
-      };
-      const run = ++mcpStatusRun;
-      await refreshMcpStatus(run);
-    } catch (e) {
-      mcpErrorByHost = {
-        ...mcpErrorByHost,
-        [hostId]: e instanceof Error ? e.message : String(e),
-      };
-    } finally {
-      mcpBusyByHost = { ...mcpBusyByHost, [hostId]: null };
-    }
-  }
-
-  function formatMcpConfiguredAt(ms: number | null | undefined): string {
-    if (!ms) return "Never";
-    return new Date(ms).toLocaleString();
-  }
-
-  async function browseWorktreeBase() {
-    const selected = await open({ directory: true, title: "Select Worktree Base Directory" });
-    if (selected) updateSetting("worktreeBasePath", selected as string);
-  }
-
-  async function browseDefaultProject() {
-    const selected = await open({ directory: true, title: "Select Default Project Directory" });
-    if (selected) updateSetting("defaultProjectPath", selected as string);
-  }
-
-  function sanitizePathInput(path: string): string {
-    return path.trim();
-  }
-
-  function addRepoRoot(path: string) {
-    const nextPath = sanitizePathInput(path);
-    if (!nextPath) return;
-    const existing = $settings.repoRoots ?? [];
-    if (existing.includes(nextPath)) {
-      repoRootDraft = "";
-      return;
-    }
-    updateSetting("repoRoots", [...existing, nextPath]);
-    repoRootDraft = "";
-  }
-
-  function removeRepoRoot(path: string) {
-    updateSetting(
-      "repoRoots",
-      ($settings.repoRoots ?? []).filter((root) => root !== path),
-    );
-  }
-
-  // Pick a repo path to use for the Settings worktree-base-path preview.
-  // Falls back to an illustrative placeholder when the user hasn't
-  // configured any repo roots or a default project yet — the preview is
-  // purely informational, so a made-up path is fine.
-  const PREVIEW_FALLBACK = "/Users/you/src/my-project";
-  let previewText = $state<string>("");
-  function previewRepoPath(): string {
-    const roots = $settings.repoRoots ?? [];
-    if (roots.length > 0) return roots[0];
-    if ($settings.defaultProjectPath) return $settings.defaultProjectPath;
-    return PREVIEW_FALLBACK;
-  }
-  $effect(() => {
-    const tpl = $settings.worktreeBasePath ?? "";
-    const repo = previewRepoPath();
-    // Cancel-on-stale guard: only keep the latest preview result.
-    let stale = false;
-    commands
-      .cmdPreviewWorktreeBase(tpl, repo)
-      .then((r) => { if (!stale) previewText = r; })
-      .catch(() => { if (!stale) previewText = ""; });
-    return () => { stale = true; };
-  });
-
-  function setCleanupMode(mode: WorktreeCleanupMode) {
-    updateSetting("worktreeCleanupOnClose", mode);
-    // Keep the legacy boolean in sync so any older readers (settings files,
-    // pre-migration code) still agree with the frontend.
-    updateSetting("cleanupWorktreesOnClose", mode === "always");
-  }
-
-  function setDefaultBase(mode: WorktreeDefaultBase) {
-    updateSetting("worktreeDefaultBase", mode);
-  }
-
-  function setPaneCloseMode(mode: OnPaneCloseMode) {
-    updateSetting("onPaneClose", mode);
-  }
-
-  async function browseAndAddRepoRoot() {
-    const selected = await open({ directory: true, title: "Select Repo Root Directory" });
-    if (selected) addRepoRoot(selected as string);
-  }
-
 </script>
 
 <svelte:window onkeydown={visible ? handleKey : undefined} />
@@ -542,931 +110,71 @@
     onkeydown={handleKey}
     tabindex="-1"
   >
-        <!-- Sidebar -->
-        <aside class="flex w-[180px] shrink-0 flex-col border-r border-hairline bg-bg-surface/30 py-3">
-        <div class="flex items-center gap-2 px-3 pb-2">
+    <aside class="flex w-[180px] shrink-0 flex-col border-r border-hairline bg-bg-surface/30 py-3">
+      <div class="flex items-center gap-2 px-3 pb-2">
+        <button
+          aria-label="Close settings"
+          class="cursor-pointer rounded border border-transparent bg-transparent p-1 text-text-muted transition-colors hover:border-border-subtle hover:bg-bg-hover hover:text-text-primary focus-visible:outline-none focus-visible:ring-1 focus-visible:ring-accent-dim/50"
+          onclick={onclose}
+        >
+          <X size={14} />
+        </button>
+        <div class="text-[11px] font-semibold uppercase tracking-widest text-text-muted">Preferences</div>
+      </div>
+      <nav class="flex flex-col gap-0.5 px-2">
+        {#each CATEGORIES as cat}
+          {@const Icon = cat.icon}
           <button
-            aria-label="Close settings"
-            class="cursor-pointer rounded border border-transparent bg-transparent p-1 text-text-muted transition-colors hover:border-border-subtle hover:bg-bg-hover hover:text-text-primary focus-visible:outline-none focus-visible:ring-1 focus-visible:ring-accent-dim/50"
-            onclick={onclose}
+            class="flex cursor-pointer items-center gap-2 rounded-md px-2 py-1.5 text-left text-[13px] transition-colors
+              {selected === cat.id
+                ? 'bg-accent-dim text-text-primary'
+                : 'text-text-secondary hover:bg-bg-hover'}"
+            onclick={() => (selected = cat.id)}
           >
-            <X size={14} />
+            <Icon size={14} />
+            <span>{cat.label}</span>
           </button>
-          <div class="text-[11px] font-semibold uppercase tracking-widest text-text-muted">Preferences</div>
-        </div>
-        <nav class="flex flex-col gap-0.5 px-2">
-          {#each CATEGORIES as cat}
-            {@const Icon = cat.icon}
-            <button
-              class="flex cursor-pointer items-center gap-2 rounded-md px-2 py-1.5 text-left text-[13px] transition-colors
-                {selected === cat.id
-                  ? 'bg-accent-dim text-text-primary'
-                  : 'text-text-secondary hover:bg-bg-hover'}"
-              onclick={() => (selected = cat.id)}
-            >
-              <Icon size={14} />
-              <span>{cat.label}</span>
-            </button>
-          {/each}
-        </nav>
-      </aside>
+        {/each}
+      </nav>
+    </aside>
 
-      <!-- Detail pane -->
-      <div class="flex min-w-0 flex-1 flex-col">
-        <div class="flex h-10 shrink-0 items-center border-b border-hairline px-4">
-          <h2 class="text-sm font-semibold tracking-tight">
-            {CATEGORIES.find((c) => c.id === selected)?.label}
-          </h2>
-        </div>
-
-        <div class="app-scrollbar flex-1 overflow-y-auto px-5 py-4">
-          {#if selected === "general"}
-            <div class="mt-3 rounded-xl border border-border-subtle bg-bg-surface/35 p-3">
-              <div class="flex items-start justify-between gap-3">
-                <div>
-                  <div class="text-[13px]">Theme</div>
-                  <div class="text-[11px] text-text-muted mt-0.5">Color preset for the app chrome. Terminal palette is configured separately under Terminal.</div>
-                </div>
-                <select
-                  class="bg-bg-deep border border-border rounded px-2 py-1 text-xs text-text-primary outline-none cursor-pointer appearance-none pr-6"
-                  value={$settings.theme}
-                  onchange={(e) => updateSetting("theme", e.currentTarget.value as typeof $settings.theme)}
-                >
-                  {#each THEME_DEFINITIONS as theme}
-                    <option value={theme.id}>{theme.label}</option>
-                  {/each}
-                </select>
-              </div>
-              <p class="mt-2 text-[11px] text-text-muted">
-                {THEME_DEFINITIONS.find((theme) => theme.id === $settings.theme)?.description}
-              </p>
-            </div>
-
-            <div class="mt-4 flex items-center justify-between py-2">
-              <span class="text-[13px]">UI font</span>
-              <input
-                class="bg-bg-deep border border-border rounded px-2 py-1 text-xs text-text-primary outline-none w-56 text-right focus:border-accent-dim"
-                value={$settings.uiFontFamily}
-                oninput={(e) => updateSetting("uiFontFamily", e.currentTarget.value)}
-              />
-            </div>
-
-            <div class="mt-4 rounded-xl border border-border-subtle bg-bg-surface/35 p-3">
-              <div class="flex items-start justify-between gap-3">
-                <div>
-                  <label for="settings-startup-target" class="text-[13px]">Open on launch</label>
-                  <div class="mt-0.5 text-[11px] text-text-muted">Choose the initial Roux surface after startup.</div>
-                </div>
-                <select
-                  id="settings-startup-target"
-                  aria-label="Open on launch"
-                  class="bg-bg-deep border border-border rounded px-2 py-1 text-xs text-text-primary outline-none cursor-pointer appearance-none pr-6"
-                  value={$settings.startupTarget ?? "restore"}
-                  onchange={(e) => updateStartupTarget(e.currentTarget.value as StartupTarget)}
-                >
-                  {#each STARTUP_TARGET_OPTIONS as option}
-                    <option value={option.id}>{option.label}</option>
-                  {/each}
-                </select>
-              </div>
-              {#if ($settings.startupTarget ?? "restore") === "externalTool"}
-                <div class="mt-3 flex items-center justify-between gap-3">
-                  <label for="settings-startup-external-tool" class="text-[13px]">Launch external tool</label>
-                  <select
-                    id="settings-startup-external-tool"
-                    aria-label="Launch external tool"
-                    class="bg-bg-deep border border-border rounded px-2 py-1 text-xs text-text-primary outline-none cursor-pointer appearance-none pr-6"
-                    value={$settings.startupExternalToolId ?? ""}
-                    onchange={(e) => updateSetting("startupExternalToolId", e.currentTarget.value || null)}
-                  >
-                    {#each globalExternalTools() as tool (tool.id)}
-                      <option value={tool.id}>{tool.name}</option>
-                    {/each}
-                  </select>
-                </div>
-                {#if globalExternalTools().length === 0}
-                  <div class="mt-2 text-[11px] text-amber">No enabled global external tools are available.</div>
-                {/if}
-              {/if}
-            </div>
-
-            <div class="mt-4 rounded-xl border border-border-subtle bg-bg-surface/35 p-3">
-              <div class="flex items-center justify-between py-1">
-                <div>
-                  <div class="text-[13px] font-semibold">Version</div>
-                  <div class="text-[11px] text-text-muted mt-0.5">{appVersion}</div>
-                </div>
-                <button
-                  class="rounded border border-border px-2.5 py-1 text-[11px] text-text-primary hover:bg-bg-hover disabled:opacity-50"
-                  disabled={$updateStatus.kind === "checking" || $updateStatus.kind === "downloading"}
-                  onclick={() => void runManualCheck()}
-                >
-                  {$updateStatus.kind === "checking" ? "Checking…" : "Check for updates"}
-                </button>
-              </div>
-
-              {#if $updateStatus.kind === "no-update"}
-                <div class="mt-2 text-[11px] text-text-secondary">You're on the latest version.</div>
-              {:else if $updateStatus.kind === "available"}
-                <div class="mt-3 rounded-lg border border-accent/30 bg-accent/5 p-3">
-                  <div class="text-[12px] font-semibold text-text-primary">Update available: {$updateStatus.version}</div>
-                  {#if $updateStatus.notes}
-                    <pre class="mt-2 max-h-40 overflow-y-auto whitespace-pre-wrap text-[11px] text-text-secondary">{$updateStatus.notes}</pre>
-                  {/if}
-                  <button
-                    class="mt-3 rounded border border-accent bg-accent-dim px-3 py-1 text-[11px] font-semibold text-text-primary hover:bg-accent/40"
-                    onclick={() => void performInstall()}
-                  >
-                    Install and restart
-                  </button>
-                </div>
-              {:else if $updateStatus.kind === "downloading"}
-                <div class="mt-3 rounded-lg border border-border-subtle bg-bg-surface/35 p-3">
-                  <div class="text-[11px] text-text-secondary">
-                    Downloading update{$updateStatus.progress !== null ? ` (${Math.round($updateStatus.progress * 100)}%)` : "…"}
-                  </div>
-                  <div class="mt-2 h-1.5 w-full overflow-hidden rounded bg-bg-deep">
-                    <div
-                      class="h-full bg-accent transition-[width] duration-200"
-                      style="width: {$updateStatus.progress !== null ? Math.round($updateStatus.progress * 100) : 30}%"
-                    ></div>
-                  </div>
-                </div>
-              {:else if $updateStatus.kind === "installed-restart-required"}
-                <div class="mt-3 rounded-lg border border-accent/30 bg-accent/5 p-3">
-                  <div class="text-[12px] font-semibold text-text-primary">Update installed</div>
-                  <div class="text-[11px] text-text-secondary mt-0.5">Quit and reopen Roux to finish.</div>
-                  <button
-                    class="mt-3 rounded border border-accent bg-accent-dim px-3 py-1 text-[11px] font-semibold text-text-primary hover:bg-accent/40"
-                    onclick={() => void quitApp()}
-                  >
-                    Quit Roux
-                  </button>
-                </div>
-              {:else if $updateStatus.kind === "error"}
-                <div class="mt-2 text-[11px] text-red">{describeError($updateStatus.reason)}</div>
-              {/if}
-
-              <div class="mt-4 flex items-center justify-between py-2">
-                <div>
-                  <div class="text-[13px]">Update channel</div>
-                  <div class="text-[11px] text-text-muted mt-0.5">Switching to Stable takes effect on the next stable release at or above your current version.</div>
-                </div>
-                <select
-                  class="bg-bg-deep border border-border rounded px-2 py-1 text-xs text-text-primary outline-none cursor-pointer appearance-none pr-6"
-                  value={$settings.updateChannel ?? "stable"}
-                  onchange={(e) => updateSetting("updateChannel", e.currentTarget.value as UpdateChannel)}
-                >
-                  <option value="stable">Stable</option>
-                  <option value="preRelease">Pre-release (Alpha)</option>
-                </select>
-              </div>
-
-              <div class="mt-4 flex items-center justify-between py-2">
-                <div>
-                  <div class="text-[13px]">Check for updates on launch</div>
-                  <div class="text-[11px] text-text-muted mt-0.5">Silently check in the background when Roux starts</div>
-                </div>
-                <button
-                  aria-label="Toggle auto-check on launch"
-                  class="w-9 h-5 rounded-full relative cursor-pointer transition-all border
-                    {($settings.updateCheckOnLaunch ?? true) ? 'bg-accent-dim border-accent' : 'bg-bg-deep border-border'}"
-                  onclick={() => updateSetting('updateCheckOnLaunch', !($settings.updateCheckOnLaunch ?? true))}
-                >
-                  <div class="w-3.5 h-3.5 rounded-full absolute top-0.5 transition-all
-                    {($settings.updateCheckOnLaunch ?? true) ? 'left-[18px] bg-accent' : 'left-0.5 bg-text-secondary'}"></div>
-                </button>
-              </div>
-            </div>
-
-            <div class="mt-4 flex items-center justify-between py-2">
-              <span class="text-[13px]">Sidebar position</span>
-              <select
-                class="bg-bg-deep border border-border rounded px-2 py-1 text-xs text-text-primary outline-none cursor-pointer appearance-none pr-6"
-                value={$sidebarLayout.railSide}
-                onchange={(e) => setRailSide(e.currentTarget.value as Side)}
-              >
-                <option value="left">Left</option>
-                <option value="right">Right</option>
-              </select>
-            </div>
-            <div class="flex items-center justify-between py-2">
-              <span class="text-[13px]">Status bar position</span>
-              <select
-                class="bg-bg-deep border border-border rounded px-2 py-1 text-xs text-text-primary outline-none cursor-pointer appearance-none pr-6"
-                value={$settings.statusBarPosition ?? "bottom"}
-                onchange={(e) => updateSetting("statusBarPosition", e.currentTarget.value as "top" | "bottom")}
-              >
-                <option value="top">Top</option>
-                <option value="bottom">Bottom</option>
-              </select>
-            </div>
-          {:else if selected === "sessions"}
-            <div class="flex items-center justify-between py-2">
-              <div>
-                <div class="text-[13px]">Confirm on close</div>
-                <div class="text-[11px] text-text-muted mt-0.5">Prompt before closing active sessions</div>
-              </div>
-              <button
-                aria-label="Toggle confirm on close"
-                class="w-9 h-5 rounded-full relative cursor-pointer transition-all border
-                  {$settings.confirmOnClose ? 'bg-accent-dim border-accent' : 'bg-bg-deep border-border'}"
-                onclick={() => updateSetting("confirmOnClose", !$settings.confirmOnClose)}
-              >
-                <div class="w-3.5 h-3.5 rounded-full absolute top-0.5 transition-all
-                  {$settings.confirmOnClose ? 'left-[18px] bg-accent' : 'left-0.5 bg-text-secondary'}"></div>
-              </button>
-            </div>
-            <div class="flex items-center justify-between py-2">
-              <div>
-                <div class="text-[13px]">Restore on launch</div>
-                <div class="text-[11px] text-text-muted mt-0.5">Show previous sessions on startup</div>
-              </div>
-              <button
-                aria-label="Toggle restore sessions on launch"
-                class="w-9 h-5 rounded-full relative cursor-pointer transition-all border
-                  {$settings.restoreSessionsOnLaunch ? 'bg-accent-dim border-accent' : 'bg-bg-deep border-border'}"
-                onclick={() => updateSetting("restoreSessionsOnLaunch", !$settings.restoreSessionsOnLaunch)}
-              >
-                <div class="w-3.5 h-3.5 rounded-full absolute top-0.5 transition-all
-                  {$settings.restoreSessionsOnLaunch ? 'left-[18px] bg-accent' : 'left-0.5 bg-text-secondary'}"></div>
-              </button>
-            </div>
-            <div class="flex items-center justify-between py-2">
-              <div>
-                <div class="text-[13px]">On pane close</div>
-                <div class="text-[11px] text-text-muted mt-0.5">Kill the terminal by default, or keep it running detached for later reconnect.</div>
-              </div>
-              <div class="flex overflow-hidden rounded border border-border bg-bg-deep">
-                {#each PANE_CLOSE_OPTIONS as opt}
-                  {@const active = ($settings.onPaneClose ?? "kill") === opt.id}
-                  <button
-                    class="px-2.5 py-1 text-[11px] cursor-pointer transition-colors
-                      {active ? 'bg-accent-dim text-text-primary' : 'text-text-secondary hover:bg-bg-hover'}"
-                    aria-pressed={active}
-                    onclick={() => setPaneCloseMode(opt.id)}
-                  >{opt.label}</button>
-                {/each}
-              </div>
-            </div>
-            <div class="flex items-center justify-between py-2">
-              <span class="text-[13px]">Default project path</span>
-              <div class="flex gap-1">
-                <input
-                  class="bg-bg-deep border border-border rounded px-2 py-1 font-mono text-xs text-text-primary outline-none w-48 text-right focus:border-accent-dim"
-                  value={$settings.defaultProjectPath ?? ""}
-                  oninput={(e) => updateSetting("defaultProjectPath", e.currentTarget.value || null)}
-                  placeholder="~/src"
-                />
-                <button
-                  class="px-2 py-1 bg-bg-elevated border border-border rounded text-text-secondary text-[10px] cursor-pointer hover:bg-bg-hover"
-                  onclick={browseDefaultProject}
-                >...</button>
-              </div>
-            </div>
-            <div class="py-2">
-              <div class="text-[13px]">Repository roots</div>
-              <div class="text-[11px] text-text-muted mt-0.5">Quick-pick sources for New Session (keeps file picker available)</div>
-              <div class="mt-2 flex gap-1">
-                <input
-                  class="bg-bg-deep border border-border rounded px-2 py-1 font-mono text-xs text-text-primary outline-none flex-1 focus:border-accent-dim"
-                  value={repoRootDraft}
-                  oninput={(e) => (repoRootDraft = e.currentTarget.value)}
-                  onkeydown={(e) => {
-                    if (e.key === "Enter") {
-                      e.preventDefault();
-                      addRepoRoot(repoRootDraft);
-                    }
-                  }}
-                  placeholder="~/src"
-                />
-                <button
-                  class="px-2 py-1 bg-bg-elevated border border-border rounded text-text-secondary text-[10px] cursor-pointer hover:bg-bg-hover"
-                  onclick={() => addRepoRoot(repoRootDraft)}
-                >Add</button>
-                <button
-                  class="px-2 py-1 bg-bg-elevated border border-border rounded text-text-secondary text-[10px] cursor-pointer hover:bg-bg-hover"
-                  onclick={browseAndAddRepoRoot}
-                >...</button>
-              </div>
-              {#if ($settings.repoRoots ?? []).length > 0}
-                <div class="mt-2 flex flex-col gap-1">
-                  {#each ($settings.repoRoots ?? []) as root (root)}
-                    <div class="flex items-center gap-2 rounded border border-border-subtle bg-bg-surface/35 px-2 py-1">
-                      <span class="font-mono text-[11px] text-text-secondary flex-1 truncate" title={root}>{root}</span>
-                      <button
-                        class="text-[10px] text-text-muted hover:text-red cursor-pointer"
-                        onclick={() => removeRepoRoot(root)}
-                      >Remove</button>
-                    </div>
-                  {/each}
-                </div>
-              {/if}
-            </div>
-            <div class="flex items-center justify-between py-2">
-              <div>
-                <div class="text-[13px]">Exclude worktrees from roots</div>
-                <div class="text-[11px] text-text-muted mt-0.5">Hide linked git worktrees from root-folder quick-pick results</div>
-              </div>
-              <button
-                aria-label="Toggle excluding worktrees from root discovery"
-                class="w-9 h-5 rounded-full relative cursor-pointer transition-all border
-                  {($settings.excludeWorktreesFromRepoRoots ?? true) ? 'bg-accent-dim border-accent' : 'bg-bg-deep border-border'}"
-                onclick={() => updateSetting("excludeWorktreesFromRepoRoots", !($settings.excludeWorktreesFromRepoRoots ?? true))}
-              >
-                <div class="w-3.5 h-3.5 rounded-full absolute top-0.5 transition-all
-                  {($settings.excludeWorktreesFromRepoRoots ?? true) ? 'left-[18px] bg-accent' : 'left-0.5 bg-text-secondary'}"></div>
-              </button>
-            </div>
-            <div class="py-2">
-              <div class="flex items-center justify-between">
-                <div>
-                  <div class="text-[13px]">Worktree base path</div>
-                  <div class="text-[11px] text-text-muted mt-0.5">
-                    Where to create new worktrees. Supports
-                    <code class="font-mono">{'{project_dir}'}</code>,
-                    <code class="font-mono">{'{git_root}'}</code>,
-                    <code class="font-mono">{'{project_name}'}</code>,
-                    <code class="font-mono">{'{home}'}</code>.
-                  </div>
-                </div>
-                <div class="flex gap-1">
-                  <input
-                    class="bg-bg-deep border border-border rounded px-2 py-1 font-mono text-xs text-text-primary outline-none w-64 text-right focus:border-accent-dim"
-                    value={$settings.worktreeBasePath ?? ""}
-                    oninput={(e) => updateSetting("worktreeBasePath", e.currentTarget.value || null)}
-                    placeholder="{'{project_dir}'}/.worktrees"
-                  />
-                  <button
-                    class="px-2 py-1 bg-bg-elevated border border-border rounded text-text-secondary text-[10px] cursor-pointer hover:bg-bg-hover"
-                    onclick={browseWorktreeBase}
-                  >...</button>
-                </div>
-              </div>
-              {#if previewText}
-                <div class="mt-1.5 text-[11px] text-text-muted font-mono truncate" title={previewText}>
-                  → {previewText}
-                </div>
-              {/if}
-            </div>
-            <div class="flex items-center justify-between py-2">
-              <div>
-                <div class="text-[13px]">On session close</div>
-                <div class="text-[11px] text-text-muted mt-0.5">What to do with the session's worktree</div>
-              </div>
-              <div class="flex rounded border border-border bg-bg-deep overflow-hidden">
-                {#each [
-                  { id: "never", label: "Keep" },
-                  { id: "prompt", label: "Ask" },
-                  { id: "always", label: "Remove" },
-                ] as const as opt}
-                  {@const active = ($settings.worktreeCleanupOnClose ?? "prompt") === opt.id}
-                  <button
-                    class="px-2.5 py-1 text-[11px] cursor-pointer transition-colors
-                      {active ? 'bg-accent-dim text-text-primary' : 'text-text-secondary hover:bg-bg-hover'}"
-                    onclick={() => setCleanupMode(opt.id)}
-                  >{opt.label}</button>
-                {/each}
-              </div>
-            </div>
-            <div class="flex items-center justify-between py-2">
-              <div>
-                <div class="text-[13px]">New Worktree default</div>
-                <div class="text-[11px] text-text-muted mt-0.5">Default starting point for new worktree branches — applies to the New Session dialog and the "New Worktree" context-menu click. Hover / command palette always expose all three.</div>
-              </div>
-              <select
-                class="bg-bg-deep border border-border rounded px-2 py-1 text-xs text-text-primary outline-none cursor-pointer appearance-none pr-6"
-                value={$settings.worktreeDefaultBase ?? "currentBranch"}
-                onchange={(e) => setDefaultBase(e.currentTarget.value as WorktreeDefaultBase)}
-              >
-                <option value="currentBranch">Current branch</option>
-                <option value="main">main</option>
-                <option value="originMain">origin/main</option>
-              </select>
-            </div>
-          {:else if selected === "agents"}
-            <SettingsAgentsSection />
-          {:else if selected === "kanban"}
-            <SettingsKanbanSection />
-          {:else if selected === "terminal"}
-            <SettingsTerminalSection />
-          {:else if selected === "notes"}
-            <SettingsNotesSection />
-          {:else if selected === "integrations" || selected === "externalTools"}
-            {#if selected === "integrations"}
-            <div class="mt-3 rounded-xl border border-border-subtle bg-bg-surface/35 p-3">
-              <div class="flex items-center justify-between">
-                <div class="text-[13px] font-semibold">Roux MCP</div>
-                {#if mcpStatus?.enabled}
-                  <span class="rounded bg-green/10 px-1.5 py-0.5 text-[10px] font-semibold uppercase tracking-wider text-green">enabled</span>
-                {:else}
-                  <span class="rounded bg-bg-active px-1.5 py-0.5 text-[10px] font-semibold uppercase tracking-wider text-text-muted">off</span>
-                {/if}
-              </div>
-              <div class="mt-0.5 text-[11px] text-text-muted">
-                Lets MCP clients launch <code class="font-mono">roux mcp</code> and use Roux sessions, panes, and notes through the running app.
-              </div>
-              <div class="mt-3 flex items-center justify-between py-1">
-                <div>
-                  <div class="text-[13px]">Enable Roux MCP</div>
-                  <div class="mt-0.5 text-[11px] text-text-muted">Safe action tools are available by default; destructive tools are not exposed.</div>
-                </div>
-                <button
-                  aria-label="Toggle Roux MCP"
-                  class="w-9 h-5 rounded-full relative cursor-pointer transition-all border
-                    {($settings.mcpEnabled ?? false) ? 'bg-accent-dim border-accent' : 'bg-bg-deep border-border'}"
-                  onclick={() => updateSetting("mcpEnabled", !($settings.mcpEnabled ?? false))}
-                >
-                  <div class="w-3.5 h-3.5 rounded-full absolute top-0.5 transition-all
-                    {($settings.mcpEnabled ?? false) ? 'left-[18px] bg-accent' : 'left-0.5 bg-text-secondary'}"></div>
-                </button>
-              </div>
-              {#if mcpStatus}
-                <div class="mt-2 grid grid-cols-2 gap-2 text-[11px]">
-                  <div class="rounded border border-border-subtle bg-bg-deep/70 px-2 py-1">
-                    <div class="text-text-muted">CLI</div>
-                    <div class={mcpStatus.cliInstalled && mcpStatus.cliCurrent ? "text-green" : "text-red"}>
-                      {mcpStatus.cliInstalled ? (mcpStatus.cliCurrent ? "current" : "stale") : "missing"}
-                    </div>
-                  </div>
-                  <div class="rounded border border-border-subtle bg-bg-deep/70 px-2 py-1">
-                    <div class="text-text-muted">Server path</div>
-                    <div class="truncate font-mono text-text-secondary" title={mcpStatus.cliPath}>{mcpStatus.cliPath}</div>
-                  </div>
-                  <div class="col-span-2 rounded border border-border-subtle bg-bg-deep/70 px-2 py-1">
-                    <div class="text-text-muted">Last config update</div>
-                    <div class="text-text-secondary">
-                      {mcpStatus.lastConfiguredAtMs ? formatMcpConfiguredAt(mcpStatus.lastConfiguredAtMs) : "Never"}
-                      {mcpStatus.lastConfiguredHost ? ` · ${mcpStatus.lastConfiguredHost}` : ""}
-                    </div>
-                  </div>
-                </div>
-              {/if}
-              {#if mcpStatus}
-                {#each mcpStatus.hosts as host (host.id)}
-                  {@const hostId = host.id as McpHostIdT}
-                  {@const busy = mcpBusyByHost[hostId] ?? null}
-                  {@const message = mcpMessageByHost[hostId] ?? null}
-                  {@const error = mcpErrorByHost[hostId] ?? null}
-                  {@const preview = mcpPreviewByHost[hostId] ?? null}
-                  <div class="mt-3 rounded border border-border-subtle bg-bg-deep/60 p-2">
-                    <div class="flex items-center justify-between gap-2">
-                      <div>
-                        <div class="text-[12px] font-medium">{host.label}</div>
-                        <div class="mt-0.5 max-w-[22rem] truncate font-mono text-[10px] text-text-muted" title={host.configPath ?? ""}>
-                          {host.configPath ?? "Config path unavailable"}
-                        </div>
-                      </div>
-                      {#if host.configured}
-                        <span class="rounded bg-green/10 px-1.5 py-0.5 text-[10px] font-semibold uppercase tracking-wider text-green">configured</span>
-                      {:else if host.error}
-                        <span class="rounded bg-red/10 px-1.5 py-0.5 text-[10px] font-semibold uppercase tracking-wider text-red">needs attention</span>
-                      {:else}
-                        <span class="rounded bg-bg-active px-1.5 py-0.5 text-[10px] font-semibold uppercase tracking-wider text-text-muted">not configured</span>
-                      {/if}
-                    </div>
-                    <div class="mt-2 flex gap-1">
-                      <button
-                        class="rounded border border-border bg-bg-elevated px-2 py-1 text-[11px] text-text-secondary hover:bg-bg-hover disabled:cursor-not-allowed disabled:opacity-40"
-                        disabled={!($settings.mcpEnabled ?? false) || busy !== null}
-                        onclick={() => previewMcpHostConfig(hostId, host.label)}
-                      >{busy === "preview" ? "Previewing" : "Preview"}</button>
-                      <button
-                        class="rounded border border-accent-dim bg-accent/15 px-2 py-1 text-[11px] text-text-primary hover:bg-accent/25 disabled:cursor-not-allowed disabled:opacity-40"
-                        disabled={!($settings.mcpEnabled ?? false) || busy !== null}
-                        onclick={() => configureMcpHost(hostId, host.label)}
-                      >{busy === "configure" ? "Adding…" : `Add to ${host.label}`}</button>
-                    </div>
-                    {#if message}
-                      <div class="mt-2 text-[11px] text-green">{message}</div>
-                    {/if}
-                    {#if error || host.error}
-                      <div class="mt-2 text-[11px] text-red">{error ?? host.error}</div>
-                    {/if}
-                    {#if preview}
-                      <div class="mt-2 rounded border border-border-subtle bg-bg-deep/70 p-2">
-                        <div class="mb-1 text-[10px] uppercase tracking-wider text-text-muted">Roux entry</div>
-                        <pre class="app-scrollbar max-h-28 overflow-auto whitespace-pre-wrap break-words font-mono text-[10px] text-text-secondary">{preview.nextEntryJson}</pre>
-                      </div>
-                    {/if}
-                  </div>
-                {/each}
-              {/if}
-            </div>
-            {/if}
-
-            {#if selected === "externalTools"}
-            <div class="mt-3 rounded-xl border border-border-subtle bg-bg-surface/35 p-3">
-              <div class="flex items-center justify-between gap-2">
-                <div>
-                  <div class="text-[13px] font-semibold">External Tools</div>
-                  <div class="mt-0.5 text-[11px] text-text-muted">
-                    Launch terminal and local web tools into the main view.
-                  </div>
-                </div>
-                <div class="flex gap-1">
-                  <button
-                    type="button"
-                    class="rounded border border-border bg-bg-elevated px-2 py-1 text-[11px] text-text-secondary hover:bg-bg-hover"
-                    onclick={() => addExternalTool("terminal")}
-                  >Add Terminal</button>
-                  <button
-                    type="button"
-                    class="rounded border border-border bg-bg-elevated px-2 py-1 text-[11px] text-text-secondary hover:bg-bg-hover"
-                    onclick={() => addExternalTool("web")}
-                  >Add Web</button>
-                </div>
-              </div>
-
-              <div class="mt-3 flex flex-col gap-2">
-                {#each externalTools() as tool (externalToolRowKey(tool.id))}
-                  {@const expanded = expandedExternalToolId === tool.id}
-                  {@const preview = externalToolPreviewById[tool.id]}
-                  <div class="rounded border border-border-subtle bg-bg-deep/60 p-2">
-                    <div class="flex items-center gap-2">
-                      <button
-                        type="button"
-                        class="min-w-0 flex-1 truncate text-left text-[12px] font-medium text-text-primary"
-                        onclick={() => (expandedExternalToolId = expanded ? null : tool.id)}
-                      >
-                        {tool.name}
-                      </button>
-                      <span class="rounded bg-bg-active px-1.5 py-0.5 text-[10px] uppercase tracking-wider text-text-muted">
-                        {tool.surface ?? "terminal"}
-                      </span>
-                      <label class="flex items-center gap-1 text-[11px] text-text-muted">
-                        <input
-                          type="checkbox"
-                          class="h-3 w-3 accent-accent"
-                          checked={tool.enabled !== false}
-                          onchange={(e) => updateExternalTool(tool.id, { enabled: e.currentTarget.checked })}
-                        />
-                        Enabled
-                      </label>
-                      <button
-                        type="button"
-                        class="rounded border border-border-subtle bg-bg-elevated px-2 py-0.5 text-[10px] text-text-secondary hover:text-red"
-                        onclick={() => removeExternalTool(tool.id)}
-                      >Remove</button>
-                    </div>
-
-                    {#if expanded}
-                      <div class="mt-3 grid gap-2">
-                        <label class="grid gap-1 text-[11px] text-text-muted">
-                          <span>Name</span>
-                          <input
-                            class="rounded border border-border bg-bg-deep px-2 py-1 text-xs text-text-primary outline-none focus:border-accent-dim"
-                            value={tool.name}
-                            oninput={(e) => updateExternalTool(tool.id, { name: e.currentTarget.value })}
-                          />
-                        </label>
-                        <div class="grid gap-2 md:grid-cols-[1fr_1fr]">
-                          <label class="grid gap-1 text-[11px] text-text-muted">
-                            <span>ID</span>
-                            <input
-                              class="rounded border border-border bg-bg-deep px-2 py-1 font-mono text-xs text-text-primary outline-none focus:border-accent-dim"
-                              value={tool.id}
-                              oninput={(e) => updateExternalTool(tool.id, { id: e.currentTarget.value })}
-                            />
-                          </label>
-                          <label class="grid gap-1 text-[11px] text-text-muted">
-                            <span>Surface</span>
-                            <input
-                              class="rounded border border-border bg-bg-deep px-2 py-1 text-xs text-text-secondary"
-                              value={tool.surface ?? "terminal"}
-                              readonly
-                            />
-                          </label>
-                        </div>
-                        <label class="grid gap-1 text-[11px] text-text-muted">
-                          <span>Command template{(tool.surface ?? "terminal") === "web" ? " (optional)" : ""}</span>
-                          <textarea
-                            class="min-h-16 rounded border border-border bg-bg-deep px-2 py-1 font-mono text-xs text-text-primary outline-none focus:border-accent-dim"
-                            value={tool.commandTemplate}
-                            oninput={(e) => updateExternalTool(tool.id, { commandTemplate: e.currentTarget.value })}
-                          ></textarea>
-                        </label>
-                        <label class="grid gap-1 text-[11px] text-text-muted">
-                          <span>CWD template</span>
-                          <input
-                            class="rounded border border-border bg-bg-deep px-2 py-1 font-mono text-xs text-text-primary outline-none focus:border-accent-dim"
-                            value={tool.cwdTemplate ?? ""}
-                            oninput={(e) => updateExternalTool(tool.id, { cwdTemplate: e.currentTarget.value })}
-                          />
-                        </label>
-                        {#if (tool.surface ?? "terminal") === "web"}
-                          <div class="grid gap-1 text-[11px] text-text-muted">
-                            <span>Embedder</span>
-                            <div class="inline-flex w-fit overflow-hidden rounded border border-border bg-bg-deep">
-                              {#each [{ value: "iframe", label: "Iframe" }, { value: "webview", label: "Webview" }] as option}
-                                <button
-                                  type="button"
-                                  class="px-2 py-1 text-[11px] transition-colors {tool.webEmbedder === option.value ? 'bg-bg-active text-text-primary' : 'text-text-muted hover:bg-bg-hover hover:text-text-secondary'}"
-                                  onclick={() => updateExternalTool(tool.id, {
-                                    webEmbedder: option.value as ExternalToolWebEmbedder,
-                                    keepWebviewAlive: option.value === "webview" ? (tool.keepWebviewAlive ?? false) : false,
-                                  })}
-                                >
-                                  {option.label}
-                                </button>
-                              {/each}
-                            </div>
-                          </div>
-                          {#if tool.webEmbedder === "webview"}
-                            <label class="flex items-center gap-2 text-[11px] text-text-secondary">
-                              <input
-                                type="checkbox"
-                                class="h-3 w-3 accent-accent"
-                                checked={tool.keepWebviewAlive === true}
-                                onchange={(e) => updateExternalTool(tool.id, { keepWebviewAlive: e.currentTarget.checked })}
-                              />
-                              Keep webview active
-                            </label>
-                          {/if}
-                          <div class="grid gap-2 md:grid-cols-[1fr_120px]">
-                            <label class="grid gap-1 text-[11px] text-text-muted">
-                              <span>URL template</span>
-                              <input
-                                class="rounded border border-border bg-bg-deep px-2 py-1 font-mono text-xs text-text-primary outline-none focus:border-accent-dim"
-                                value={tool.urlTemplate ?? ""}
-                                oninput={(e) => updateExternalTool(tool.id, { urlTemplate: e.currentTarget.value || null })}
-                              />
-                            </label>
-                            <label class="grid gap-1 text-[11px] text-text-muted">
-                              <span>Preferred port</span>
-                              <input
-                                type="number"
-                                min="1"
-                                max="65535"
-                                class="rounded border border-border bg-bg-deep px-2 py-1 font-mono text-xs text-text-primary outline-none focus:border-accent-dim"
-                                value={tool.preferredPort ?? ""}
-                                oninput={(e) => updateExternalTool(tool.id, { preferredPort: preferredPortFromInput(e.currentTarget.value) })}
-                              />
-                            </label>
-                          </div>
-                        {/if}
-                        <label class="flex items-center gap-2 text-[11px] text-text-secondary">
-                          <input
-                            type="checkbox"
-                            class="h-3 w-3 accent-accent"
-                            checked={tool.requiresSession ?? false}
-                            onchange={(e) => updateExternalTool(tool.id, { requiresSession: e.currentTarget.checked })}
-                          />
-                          Requires active session
-                        </label>
-                        <div class="flex items-center gap-2">
-                          <button
-                            type="button"
-                            class="rounded border border-border bg-bg-elevated px-2 py-1 text-[11px] text-text-secondary hover:bg-bg-hover disabled:opacity-40"
-                            disabled={preview?.loading}
-                            onclick={() => void previewTool(tool)}
-                          >{preview?.loading ? "Previewing" : "Preview Render"}</button>
-                          {#if tool.requiresSession && !$activeSession}
-                            <span class="text-[11px] text-amber">Preview needs an active session.</span>
-                          {/if}
-                        </div>
-                        {#if preview?.error}
-                          <div class="rounded border border-red/25 bg-red/10 p-2 text-[11px] text-red">
-                            {preview.error}
-                          </div>
-                        {:else if preview?.rendered}
-                          <div class="grid gap-1 rounded border border-border-subtle bg-bg-deep/70 p-2 font-mono text-[10px] text-text-secondary">
-                            <div><span class="text-text-muted">cmd</span> {preview.rendered.command}</div>
-                            <div><span class="text-text-muted">cwd</span> {preview.rendered.cwd}</div>
-                            {#if preview.rendered.url}
-                              <div><span class="text-text-muted">url</span> {preview.rendered.url}</div>
-                            {/if}
-                          </div>
-                        {/if}
-                      </div>
-                    {/if}
-                  </div>
-                {/each}
-              </div>
-            </div>
-            {/if}
-
-            {#if selected === "integrations"}
-            <div class="mt-3 rounded-xl border border-border-subtle bg-bg-surface/35 p-3">
-              <div class="flex items-center justify-between">
-                <div class="text-[13px] font-semibold">GitHub CLI</div>
-                {#if ghDetection?.binaryPath}
-                  <span
-                    class="rounded bg-green/10 px-1.5 py-0.5 text-[10px] font-semibold uppercase tracking-wider text-green"
-                    >detected{ghDetection.version ? ` ${ghDetection.version}` : ""}</span
-                  >
-                {:else if ghDetection !== null}
-                  <span
-                    class="rounded bg-red/10 px-1.5 py-0.5 text-[10px] font-semibold uppercase tracking-wider text-red"
-                    >missing</span
-                  >
-                {/if}
-              </div>
-              <div class="mt-0.5 text-[11px] text-text-muted">
-                Used for "Session from PR" and PR watches. Roux auto-detects
-                <code class="font-mono">gh</code> via your login shell's PATH (including fish). Set this only if
-                auto-detection misses your install — paste the output of <code class="font-mono">which gh</code>.
-              </div>
-              {#if ghDetection?.binaryPath}
-                <div class="mt-2 font-mono text-[10px] text-text-muted">
-                  {ghDetection.binaryPath}
-                </div>
-              {/if}
-              <div class="mt-3 flex items-center justify-between gap-2">
-                <span class="text-[13px]">Binary path</span>
-                <div class="flex gap-1">
-                  <input
-                    class="bg-bg-deep border border-border rounded px-2 py-1 font-mono text-xs text-text-primary outline-none w-64 text-right focus:border-accent-dim"
-                    value={$settings.ghBinaryPath ?? ""}
-                    oninput={(e) => updateSetting("ghBinaryPath", e.currentTarget.value || null)}
-                    placeholder="/opt/homebrew/bin/gh"
-                  />
-                  <button
-                    class="px-2 py-1 bg-bg-elevated border border-border rounded text-text-secondary text-[10px] cursor-pointer hover:bg-bg-hover"
-                    onclick={browseGhBinary}
-                  >...</button>
-                </div>
-              </div>
-              <div class="mt-3 flex items-center justify-between py-1">
-                <div>
-                  <div class="text-[13px]">Look up PR for session branch</div>
-                  <div class="text-[11px] text-text-muted mt-0.5">
-                    Run <code class="font-mono">gh pr list --head &lt;branch&gt;</code> on session activation
-                    so the status bar can show a clickable PR link.
-                  </div>
-                </div>
-                <button
-                  aria-label="Toggle session PR lookup"
-                  class="w-9 h-5 rounded-full relative cursor-pointer transition-all border
-                    {$settings.autoLookupSessionPr !== false ? 'bg-accent-dim border-accent' : 'bg-bg-deep border-border'}"
-                  onclick={() => updateSetting("autoLookupSessionPr", $settings.autoLookupSessionPr === false)}
-                >
-                  <div class="w-3.5 h-3.5 rounded-full absolute top-0.5 transition-all
-                    {$settings.autoLookupSessionPr !== false ? 'left-[18px] bg-accent' : 'left-0.5 bg-text-secondary'}"></div>
-                </button>
-              </div>
-              <div class="flex items-center justify-between py-1">
-                <div>
-                  <div class="text-[13px]">Auto-create PR watch for sessions</div>
-                  <div class="text-[11px] text-text-muted mt-0.5">
-                    When the lookup finds a PR, create a session-scoped GitHub PR watch
-                    automatically. Requires the lookup above.
-                  </div>
-                </div>
-                <button
-                  aria-label="Toggle auto PR watch"
-                  disabled={$settings.autoLookupSessionPr === false}
-                  class="w-9 h-5 rounded-full relative cursor-pointer transition-all border
-                    {$settings.autoWatchSessionPr ? 'bg-accent-dim border-accent' : 'bg-bg-deep border-border'}
-                    {$settings.autoLookupSessionPr === false ? 'opacity-40 cursor-not-allowed' : ''}"
-                  onclick={() => updateSetting("autoWatchSessionPr", !$settings.autoWatchSessionPr)}
-                >
-                  <div class="w-3.5 h-3.5 rounded-full absolute top-0.5 transition-all
-                    {$settings.autoWatchSessionPr ? 'left-[18px] bg-accent' : 'left-0.5 bg-text-secondary'}"></div>
-                </button>
-              </div>
-            </div>
-
-            <div class="mt-3 rounded-xl border border-border-subtle bg-bg-surface/35 p-3">
-              <div class="flex items-center justify-between">
-                <div class="text-[13px] font-semibold">Git</div>
-                {#if gitDetection?.binaryPath}
-                  <span
-                    class="rounded bg-green/10 px-1.5 py-0.5 text-[10px] font-semibold uppercase tracking-wider text-green"
-                    >detected{gitDetection.version ? ` ${gitDetection.version}` : ""}</span
-                  >
-                {:else if gitDetection !== null}
-                  <span
-                    class="rounded bg-red/10 px-1.5 py-0.5 text-[10px] font-semibold uppercase tracking-wider text-red"
-                    >missing</span
-                  >
-                {/if}
-              </div>
-              <div class="mt-0.5 text-[11px] text-text-muted">
-                Used for git-backed Library sources. Roux checks this override,
-                then <code class="font-mono">ROUX_GIT</code>, then your login shell's PATH, then the app PATH.
-                Set this only if auto-detection misses your install.
-              </div>
-              {#if gitDetection?.binaryPath}
-                <div class="mt-2 font-mono text-[10px] text-text-muted">
-                  {gitDetection.binaryPath}
-                </div>
-              {/if}
-              <div class="mt-3 flex items-center justify-between gap-2">
-                <span class="text-[13px]">Binary path</span>
-                <div class="flex gap-1">
-                  <input
-                    class="bg-bg-deep border border-border rounded px-2 py-1 font-mono text-xs text-text-primary outline-none w-64 text-right focus:border-accent-dim"
-                    value={$settings.gitBinaryPath ?? ""}
-                    oninput={(e) => updateSetting("gitBinaryPath", e.currentTarget.value || null)}
-                    placeholder="/opt/homebrew/bin/git"
-                  />
-                  <button
-                    class="px-2 py-1 bg-bg-elevated border border-border rounded text-text-secondary text-[10px] cursor-pointer hover:bg-bg-hover"
-                    onclick={browseGitBinary}
-                  >...</button>
-                </div>
-              </div>
-            </div>
-
-            <div class="mt-3 rounded-xl border border-border-subtle bg-bg-surface/35 p-3">
-              <div class="flex items-center justify-between">
-                <div class="text-[13px] font-semibold">Worktrunk (wt)</div>
-                {#if worktrunkDetection?.binaryPath}
-                  <span
-                    data-testid="worktrunk-detected-badge"
-                    class="rounded bg-green/10 px-1.5 py-0.5 text-[10px] font-semibold uppercase tracking-wider text-green"
-                    >detected{worktrunkDetection.version
-                      ? ` ${worktrunkDetection.version}`
-                      : ""}</span
-                  >
-                {:else}
-                  <span
-                    data-testid="worktrunk-not-detected-badge"
-                    class="rounded bg-bg-active px-1.5 py-0.5 text-[10px] font-semibold uppercase tracking-wider text-text-muted"
-                    >not detected</span
-                  >
-                {/if}
-              </div>
-              <div class="mt-0.5 text-[11px] text-text-muted">
-                When available, Roux enriches the New Session worktree picker
-                with <code class="font-mono">wt</code>'s richer metadata (dirty
-                state, ahead/behind, locked/prunable, current/previous). Opt-in
-                — no regression for users without <code class="font-mono">wt</code>.
-                Set the path only if auto-detection misses your install.
-              </div>
-              {#if worktrunkDetection?.binaryPath}
-                <div class="mt-2 font-mono text-[10px] text-text-muted">
-                  {worktrunkDetection.binaryPath}
-                </div>
-              {/if}
-              <div class="mt-3 flex items-center justify-between gap-2">
-                <span class="text-[13px]">Binary path</span>
-                <div class="flex gap-1">
-                  <input
-                    class="bg-bg-deep border border-border rounded px-2 py-1 font-mono text-xs text-text-primary outline-none w-64 text-right focus:border-accent-dim"
-                    value={$settings.worktrunkBinaryPath ?? ""}
-                    oninput={(e) =>
-                      updateSetting(
-                        "worktrunkBinaryPath",
-                        e.currentTarget.value || null
-                      )}
-                    placeholder="/opt/homebrew/bin/wt"
-                  />
-                  <button
-                    class="px-2 py-1 bg-bg-elevated border border-border rounded text-text-secondary text-[10px] cursor-pointer hover:bg-bg-hover"
-                    onclick={browseWorktrunkBinary}>...</button
-                  >
-                </div>
-              </div>
-              <div class="mt-3 flex items-center justify-between gap-2">
-                <div>
-                  <div class="text-[13px]">Worktree provider</div>
-                  <div class="mt-0.5 text-[11px] text-text-muted">
-                    How Roux creates new worktrees. Auto uses <code
-                      class="font-mono">wt</code
-                    > when detected and falls back to git otherwise.
-                  </div>
-                </div>
-                <div
-                  data-testid="worktrunk-provider-selector"
-                  class="flex overflow-hidden rounded border border-border bg-bg-deep"
-                >
-                  {#each [
-                    { id: "auto", label: "Auto" },
-                    { id: "git", label: "Git" },
-                    { id: "worktrunk", label: "wt" },
-                  ] as const as opt}
-                    {@const active =
-                      ($settings.worktreeProvider ?? "auto") === opt.id}
-                    <button
-                      data-testid={`worktrunk-provider-${opt.id}`}
-                      class="cursor-pointer px-2.5 py-1 text-[11px] transition-colors
-                        {active
-                        ? 'bg-accent-dim text-text-primary'
-                        : 'text-text-secondary hover:bg-bg-hover'}"
-                      onclick={() =>
-                        updateSetting(
-                          "worktreeProvider",
-                          opt.id as WorktreeProvider
-                        )}>{opt.label}</button
-                    >
-                  {/each}
-                </div>
-              </div>
-            </div>
-            {/if}
-          {:else if selected === "notifications"}
-            <SettingsNotificationsSection />
-          {:else if selected === "keyboard"}
-            <SettingsKeyboardSection />
-          {:else if selected === "experiments"}
-            <SettingsExperimentsSection />
-          {:else if selected === "advanced"}
-            <SettingsAdvancedSection />
-          {/if}
-        </div>
+    <div class="flex min-w-0 flex-1 flex-col">
+      <div class="flex h-10 shrink-0 items-center border-b border-hairline px-4">
+        <h2 class="text-sm font-semibold tracking-tight">
+          {CATEGORIES.find((c) => c.id === selected)?.label}
+        </h2>
       </div>
 
+      <div class="app-scrollbar flex-1 overflow-y-auto px-5 py-4">
+        {#if selected === "general"}
+          <SettingsGeneralSection />
+        {:else if selected === "sessions"}
+          <SettingsSessionsSection />
+        {:else if selected === "terminal"}
+          <SettingsTerminalSection />
+        {:else if selected === "agents"}
+          <SettingsAgentsSection />
+        {:else if selected === "kanban"}
+          <SettingsKanbanSection />
+        {:else if selected === "externalTools"}
+          <SettingsExternalToolsSection
+            focusedToolId={focusedExternalToolId}
+            focusToken={externalToolFocusToken}
+          />
+        {:else if selected === "notes"}
+          <SettingsNotesSection />
+        {:else if selected === "integrations"}
+          <SettingsIntegrationsSection />
+        {:else if selected === "notifications"}
+          <SettingsNotificationsSection />
+        {:else if selected === "keyboard"}
+          <SettingsKeyboardSection />
+        {:else if selected === "experiments"}
+          <SettingsExperimentsSection />
+        {:else if selected === "advanced"}
+          <SettingsAdvancedSection />
+        {/if}
+      </div>
+    </div>
   </div>
 {/if}
